+++ src/frontend/web_application/src/modules/pi/components/MultidimensionalPi/presenter.jsx
@@ -2,10 +2,9 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import classnames from 'classnames';
 import { v1 as uuidV1 } from 'uuid';
-import PiGraph from './components/PiGraph';
-import Ratings from './components/Ratings';
-import Button from '../Button';
-import { Dropdown, withDropdownControl } from '../Dropdown';
+import PiGraph from '../PiGraph';
+import Ratings from '../Ratings';
+import { Button, Dropdown, withDropdownControl } from '../../../../components/';
 
 import './style.scss';
