+++ src/frontend/web_application/src/modules/pi/components/MultidimensionalPi/style.scss
@@ -1,6 +1,6 @@
-@import '../../styles/common';
-@import '../../styles/vendor/bootstrap_foundation-sites';
-@import '../../styles/object/o-triangle';
+@import '../../../../styles/common';
+@import '../../../../styles/vendor/bootstrap_foundation-sites';
+@import '../../../../styles/object/o-triangle';
 
 
 .m-multidimensional-pi {
