+++ src/frontend/web_application/src/modules/pi/components/PiGraph/style.scss
