+++ src/frontend/web_application/src/modules/pi/components/Ratings/style.scss
