+++ src/frontend/web_application/src/modules/pi/index.jsx
@@ -1,0 +1,1 @@
+export { default as MultidimensionalPi } from './components/MultidimensionalPi';
