+++ src/frontend/web_application/src/modules/pi/services/svg/index.spec.jsx
