+++ src/frontend/web_application/src/scenes/Contact/components/ContactProfile/index.jsx
@@ -3,8 +3,8 @@ import PropTypes from 'prop-types';
 import classnames from 'classnames';
 import { withI18n } from 'lingui-react';
 import { WithTags, getTagLabel, getCleanedTagCollection } from '../../../../modules/tags';
-import Badge from '../../../../components/Badge';
-import MultidimensionalPi from '../../../../components/MultidimensionalPi';
+import { MultidimensionalPi } from '../../../../modules/pi';
+import { Badge } from '../../../../components';
 import ContactAvatarLetter from '../../../../components/ContactAvatarLetter';
 import { formatName } from '../../../../services/contact';
 import './style.scss';
